Removed confusing loader.load behaviour, now always returns a tuple
Previously if Flat and Dark images were not loaded, a single value would be returned. This led to confusing documentation and usage. Now it always returns a tuple. If no Flat and Dark images are specified, they will be returned as None. To specify that no Flat and Dark images are expected in the code, they should be called _ for anonymous argument.