Handle COR/Tilt finding parameters as an image operation
This treats COR/Tilt finding results as an imahe operation that is
placed in the stack history.

This allows determining if a crop operation happened before or after a
COR/Tilt finding operation.

Performing crop operations before COR/TIlt finding would have previously
led to incorrect calculation of the rotation centre parameter as all
crops were assumed to have happened after the COR/Tilt finding.

This further takes into account crops on the top of the image, which
alter the rotation centre parameter due to the fact this parameter is
tied to the top of the image at the time of reconstruction.

Fixes #248