Force upload to Anaconda Cloud
This fixes the situation where if (for whatever reason) the master CI
job is ran twice with the same commit history (therefore generating the
same version identifier) the subsequent CI jobs fail on the Anaconda
upload step as they are trying to upload a version that already exists.

This will now replace whatever is uplaoded to Anaconda cloud for a given
version identifier with the version just built.