Adds do_apply_filter_sync which applies synchronously
This works around a limitation in the parallel package which
can't tolerate multiple async functions accessing the global share.

Not the best fix but I thought it the lowest risk one.

A better solution could be to instead move the proj180 handling logic
into `do_apply_filter` and remove it from `_post_filter`, but that
requires a decent refactor and is a higher risk of breaking functionality.

Fixes #711