Add ROI normalisation modes
'Preserve Max' : Normalisation is scaled such that the the maximum pixel value of the stack is equal before
	     and after the operation.
'Stack Average' : The mean value of the air region across all projections is preserved.
'Flat Field' : The mean value of the air regions in the projects is made equal to the mean value of the air
	   region in the flat field.

'Preserve Max' is equivilent to the old behaviour.