Appease mypy 0.900
pkgutil and import lib give unions and optionals that upset mypy. Here
we know what the concrete types are and so can assert them.

Also need to convince mypy that the modules have a FILTER_CLASS
attribute. This is done by creating a protocol and casting to it. This
is safe because the previous line has check the attribute exists.