Recon use local reference to self.images
self.images can be updated during the running of and async task if the user selects a new stack. This caused some race conditions in tests. Although they have not been seen in use, it makes sense to make a local reference so that the whole function acts on the same image.