Replace filter_applied signal with stack_changed
filter_applied:
We are more interested in finding out that a stack has changed, by an operation, or any thing else. So rename this to stack_changed. This will also be used to trigger redraws when needed.

recon_applied:
Was only used by the wizard. But a complete recon results a new stack, and so model_changed is already emitted. Therefore recon_applied is not needed and can be removed.