Add shared.run_compute_func
This allows operations running compute functions that give more control.

This is designed to be a more flexible replacement for shared.execute. The operation needs to provide a funtion that takes an index, list of arrays and dict of parameters. run_compute_func takes care of running the function over a range of indexes. The compute function can use whatever slicing and indexing is useful.