Refactor Handling of Stack Setting
Refactor handling of ROIs when setting a stack based on the current ROI state. If there are additional ROIs present when  is called, those ROIs are removed and the states of default ROI are reset to the reflect the change in sample. If there are no additional ROIs then then  executes according to previous method behaviour assuming no ROIs existed previously