Fix bug where CIL reconstruction was being passed the wrong COR
This bug occured because a previous commit changed the Geometry logic to accept the topmost COR (matching MI convention), converting it to the midpoint COR (matching CIL convention), rather than passing in the midpoint COR directly. However, the call to set_geometry_from_cor_tilt() wasn't updated to reflect this.