Don't try to run previews before window is fully setup
This is needed to avoid an initialisation ordering issue. Changing either the stack or the operation will trigger a preview, but both need to be set for the preview to succeed. The previous workaround was to not do a preview when there are no parameters set which is the case when the operation has not yet been set. However this is also true for operations that have no parameters, such as monitor normalisation.

Instead have a boolean flag in the view, that is set true at the end of the constructor.