Conda workflow: Cache the conda env directly
In the original version a copy of the conda environment was made and copied. This does not seem to be needed any more. This saves a lot of space

Move the caching of the env until after miniconda is set up.

Also give a fixed path, as I saw some errors using a variable in the path.